requirements.txt
# ── Backend ──
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
httpx>=0.27.0

# ── RAG ──
chromadb>=0.5.0
sentence-transformers>=3.0.0

# ── LangGraph ──
langgraph>=0.2.0

# ── Report ──
jinja2>=3.1.0
pdfkit>=1.0.0

# ── Fine-tuning (필요 시 설치) ──
# torch>=2.0.0
# transformers>=4.40.0
# peft>=0.12.0
# trl>=0.9.0
# bitsandbytes>=0.43.0
# datasets>=2.0.0
